+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.4.1] - 2024-10-18
+### Changed
+- Admin Settings page form & text
+- Settings keys
+- Migration (first, general default settings)
+
 ## [0.4.0] - 2024-10-15
 ### Added
 - Migrations: default image column to tags table